Fixing tab-completion on beginning of line
The "artifact" (which was the token used for completion) is not show any
more, also the complete crash on no matching channels or users on
tab-completion has also been fixed.

Also moved the tab_completion method to separate file, for better focus.

closing #37
